Fixes frequent freezing updates of firebase
