+++ src/components/organisms/Chatbox/Chatbox.tsx
@@ -18,11 +18,11 @@ import { WithId } from "utils/id";
 import { DEFAULT_PARTY_NAME, DEFAULT_PROFILE_IMAGE } from "settings";
 import { chatSort } from "utils/chat";
 import {
-  partygoersSelector,
   partygoersSelectorData,
   privateChatsSelector,
   venueChatsSelector,
 } from "utils/selectors";
+import { usePartygoers } from "hooks/useUsers";
 
 // Don't pull everything
 // REVISIT: only grab most recent N from server
@@ -55,7 +55,7 @@ const Chatbox: React.FunctionComponent<PropsType> = ({
 
   const { user } = useUser();
   const users = useSelector(partygoersSelectorData);
-  const userArray = useSelector(partygoersSelector) ?? [];
+  const userArray = usePartygoers();
 
   const [searchValue, setSearchValue] = useState<string>("");
   const debouncedSearch = debounce((v) => setSearchValue(v), 500);
