+++ src/components/templates/Camp/components/RoomModal.tsx
@@ -6,11 +6,7 @@ import { CampRoomData } from "types/CampRoomData";
 
 import { getCurrentEvent } from "utils/event";
 import { enterLocation } from "utils/useLocationUpdateEffect";
-import {
-  currentVenueSelector,
-  orderedVenuesSelector,
-  partygoersSelector,
-} from "utils/selectors";
+import { currentVenueSelector, orderedVenuesSelector } from "utils/selectors";
 import {
   getCurrentTimeInUnixEpochSeconds,
   ONE_MINUTE_IN_SECONDS,
@@ -25,6 +21,7 @@ import { RoomModalOngoingEvent } from "./RoomModalOngoingEvent";
 import { ScheduleItem } from "./ScheduleItem";
 
 import "components/templates/PartyMap/components/RoomModal/RoomModal.scss";
+import { usePartygoers } from "hooks/useUsers";
 
 interface RoomModalProps {
   show: boolean;
@@ -47,7 +44,7 @@ export const RoomModal: React.FC<RoomModalProps> = ({
   const venueEvents = useSelector(
     (state) => state.firestore.ordered.venueEvents
   );
-  const users = useSelector(partygoersSelector) ?? [];
+  const users = usePartygoers();
 
   const usersToDisplay = useMemo(
     () =>
